better frontmatter support, snapshot testing
This commit adds snapshot testing for error messages and removes the old
`frontmatter` namespace in favor of new options given to both
`ziggy.Parser` (the type-driven parser) and `ziggy.Ast`.

This change means that now frontmatter delimiters are part of the
Ziggy grammar and are recognized as valid tokens **if** the parser
has been instructed to look for them (they will be an error otherwise).

With this change it's now possible to have Ziggy parse directly a file
that embeds a Ziggy frontmatter (e.g. SuperMD).

Two components are still missing support for frontmatter tokens:
- LSP (and the various dedicated parsers)
- Schema (you can't check a smd file for schema compliance, for example)

Future commits will work on adding the missing support.