init lsp support for frontmatter ziggy
This commit introduces support for detecting
frontmatter framing tokens in ResilientParser and
gives the LSP the ability to be used with files
with language set to `supermd`.

This implementation is incomplete as it only properly
deals with frontmatter tokens in the top-level-struct
case and not in other cases, which causes the LSP
to occasionally try to parse outside the boundaries
of the frontmatter.