better lsp frontmatter handling
New strategy for handling frontmatter documents
based on having the LSP do the framing work.

I had to comment out an assertion that in theory
should be correct, but that trips on braceless
top level structs with a trailing comma.

This new approach only requires the LSP parser
to concern itself with correctly handling the
opening '---' token as the closing one (and any
subsequent data) is excluded from the slice of
bytes fed to the parser.