wip: parse JSON numbers natively
This PR will close #1.

Currently implements a finite state automaton that accepts valid JSON
numbers. The JSON number specification has a couple interesting
features, for example `0.0e0` is valid JSON, while `02` is not.

Note: currently numbers are not parsed at all, merely accepted or
rejected as integers / floats by the finite state automaton. A next step
would be to audit the parsing code in `std.zig.parse_float.zig` to see
to what extent it can be used for JSON numbers. I believe that the
current parsing in `std.json`, while likely not incorrect, seems like
not a great fit for this PR, since it shells out to `std.fmt`.